Insert a new exercise that will switch remote transport to Artery/TCP
This commit just copies the base cluster example into a new exercise.
The actual switching of the remote transport is done in the next
commit. The README file for the new exercise is already updated though.

Starting for the base Akka Cluster example, the Akka remote transport
will be switched from the default one to Artery TCP (introduced in Akka
2.5.11)