Add support for new 10-LED strip
Add support for 8-LED strip where LED #0 is in the most right position and
the upcoming 10-LED strip where LED #0 is on the most left position.

Selection of the strip when running the exercises is done via setting an
environment variable 'LED_STRIP_TYPE' to one of two possible values (see
reference.conf in 'common' project for details).